requirements.txt
addict==2.3.0
Cython==0.29.16
filterpy==1.4.5
lap==0.4.0
matplotlib==3.2.1
motmetrics==1.2.0
munkres==1.1.2
numba==0.48.0
numpy==1.18.2
numpy-quaternion==2020.4.24.9.29.7
nuscenes-devkit==1.1.1
open3d==0.10.0.0
opencv-python==4.2.0.34
pylint==2.5.3
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
pycocotools==2.0.2
pytest==6.1.0
seaborn==0.10.0
six==1.15.0
tensorboard==1.15.0
tensorboardX==2.0
tensorflow-gpu==1.15.0
termcolor==1.1.0
terminaltables==3.1.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.45.0
utm==0.5.0
waymo-open-dataset-tf-1-15-0==1.2.0
yapf==0.30.0
